[refactor] Separate “master” and “slave” JSON write code paths
A "master" write occurs when the root site sees a dependency change; it
needs to re-stitch the whole root menu and write it to disk.

By contrast, a "slave" write occurs on e.g. /labs, when the root menu
changes. In that case, we shall write a mirror of what we received,
without recomputing.

- Refactor to eliminate ->update(), as it was mixing a REST concern
with a cache consistency concern (and causing a gratuitous four-fold
increase in the pubsub traffic, har har)
- Perform "master" writes in the dependency-checking loop, against
the one menu that needs refreshing
- Perform "slave" writes outside of the loop i.e. regardless of
whether the local root (/labs) actually uses the menu or not